First asset up and running
