+++ .gitignore
@@ -1,2 +1,8 @@
-
+# Node 
 node_modules/
+
+# binaries
+.zip
+
+# log
+.log
