--- View/home.tpl
@@ -1,33 +1,0 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<!--
-Design by TEMPLATED
-http://templated.co
-Released for free under the Creative Commons Attribution License
-
-Name       : Saleable 
-Description: A two-column, fixed-width design with dark color scheme.
-Version    : 1.0
-Released   : 20131118
-
--->
-<html xmlns="http://www.w3.org/1999/xhtml">
-	<head>
-		<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-		<title>(:title:)</title>
-		<meta name="keywords" content="" />
-		<meta name="description" content="" />
-		<!--link href="http://fonts.googleapis.com/css?family=Source+Sans+Pro:200,300,400,600,700,900" rel="stylesheet" /-->
-		<link href="/Css/style.css" rel="stylesheet" type="text/css" media="all" />
-		<link href="/Css/fonts.css" rel="stylesheet" type="text/css" media="all" />
-<!--[if IE 6]><link href="default_ie6.css" rel="stylesheet" type="text/css" /><![endif]-->
-	</head>
-	<body>
-		<p>Welcome Guys!</p>
-		<p>Navigate to</p>
-		<ul>
-			<li><a class="button" href="./drivers">Driver's Page</a></li>
-			<li><a class="button" href="./races">Races' Page</a></li>
-			<li><a class="button" href="./admin">Admin Page</a></li>
-		</ul>
-	</body>
-</html>
